+++ R/hop.R
@@ -103,7 +103,8 @@ hop <- function(.x,
     .f,
     ...,
     .ptype = list(),
-    .constrain = FALSE
+    .constrain = FALSE,
+    .atomic = FALSE
   )
 }
 
@@ -135,7 +136,8 @@ hop_vec <- function(.x,
     .f,
     ...,
     .ptype = .ptype,
-    .constrain = TRUE
+    .constrain = TRUE,
+    .atomic = TRUE
   )
 }
 
@@ -144,15 +146,16 @@ hop_vec_simplify <- function(.x,
                              .stops,
                              .f,
                              ...) {
-  out <- hop(
+  out <- hop_impl(
     .x,
     .starts,
     .stops,
     .f,
-    ...
+    ...,
+    .ptype = list(),
+    .constrain = FALSE,
+    .atomic = TRUE
   )
-
-  check_all_size_one(out)
 
   vec_simplify(out)
 }
@@ -165,7 +168,8 @@ hop_impl <- function(.x,
                      .f,
                      ...,
                      .ptype,
-                     .constrain) {
+                     .constrain,
+                     .atomic) {
   vec_assert(.x)
 
   .f <- as_function(.f)
@@ -182,6 +186,7 @@ hop_impl <- function(.x,
     ptype = .ptype,
     env = environment(),
     type = type,
-    constrain = .constrain
+    constrain = .constrain,
+    atomic = .atomic
   )
 }