+++ R/pslide-period.R
@@ -22,8 +22,9 @@ pslide_period <- function(.l,
     .before = .before,
     .after = .after,
     .complete = .complete,
+    .ptype = list(),
     .constrain = FALSE,
-    .ptype = list()
+    .atomic = FALSE
   )
 }
 
@@ -69,8 +70,9 @@ pslide_period_vec <- function(.l,
     .before = .before,
     .after = .after,
     .complete = .complete,
+    .ptype = .ptype,
     .constrain = TRUE,
-    .ptype = .ptype
+    .atomic = TRUE
   )
 }
 
@@ -84,7 +86,7 @@ pslide_period_vec_simplify <- function(.l,
                                        .before,
                                        .after,
                                        .complete) {
-  out <- pslide_period(
+  out <- pslide_period_impl(
     .l,
     .i,
     .period,
@@ -94,10 +96,11 @@ pslide_period_vec_simplify <- function(.l,
     .origin = .origin,
     .before = .before,
     .after = .after,
-    .complete = .complete
+    .complete = .complete,
+    .ptype = list(),
+    .constrain = FALSE,
+    .atomic = TRUE
   )
-
-  check_all_size_one(out)
 
   vec_simplify(out)
 }
@@ -282,8 +285,9 @@ pslide_period_impl <- function(.l,
                                .before,
                                .after,
                                .complete,
+                               .ptype,
                                .constrain,
-                               .ptype) {
+                               .atomic) {
   check_is_list(.l)
 
   lapply(.l, vec_assert)
@@ -318,8 +322,9 @@ pslide_period_impl <- function(.l,
     before = .before,
     after = .after,
     complete = .complete,
-    constrain = .constrain,
     ptype = .ptype,
+    constrain = .constrain,
+    atomic = .atomic,
     env = environment(),
     type = type
   )