+++ R/slide-period2.R
@@ -92,8 +92,9 @@ slide_period2 <- function(.x,
     .before = .before,
     .after = .after,
     .complete = .complete,
+    .ptype = list(),
     .constrain = FALSE,
-    .ptype = list()
+    .atomic = FALSE
   )
 }
 
@@ -142,8 +143,9 @@ slide_period2_vec <- function(.x,
     .before = .before,
     .after = .after,
     .complete = .complete,
+    .ptype = .ptype,
     .constrain = TRUE,
-    .ptype = .ptype
+    .atomic = TRUE
   )
 }
 
@@ -158,7 +160,7 @@ slide_period2_vec_simplify <- function(.x,
                                        .before,
                                        .after,
                                        .complete) {
-  out <- slide_period2(
+  out <- slide_period2_impl(
     .x,
     .y,
     .i,
@@ -169,10 +171,11 @@ slide_period2_vec_simplify <- function(.x,
     .origin = .origin,
     .before = .before,
     .after = .after,
-    .complete = .complete
+    .complete = .complete,
+    .ptype = list(),
+    .constrain = FALSE,
+    .atomic = TRUE
   )
-
-  check_all_size_one(out)
 
   vec_simplify(out)
 }
@@ -370,8 +373,9 @@ slide_period2_impl <- function(.x,
                                .before,
                                .after,
                                .complete,
+                               .ptype,
                                .constrain,
-                               .ptype) {
+                               .atomic) {
   vec_assert(.x)
   vec_assert(.y)
 
@@ -396,8 +400,9 @@ slide_period2_impl <- function(.x,
     before = .before,
     after = .after,
     complete = .complete,
-    constrain = .constrain,
     ptype = .ptype,
+    constrain = .constrain,
+    atomic = .atomic,
     env = environment(),
     type = type
   )